Update classic.py
I read the original niqe paper, and it seems that prompt is not needed, and the compute code is already implemented in niqe.py. I'm not sure if there are other metrics needed in classic.py. If there are any other metrics that need to be changed, please feel free to let me know